feat: add incomemodifier and upgrade / Also upgraded to new version to fix security issue

+++ Assets/Scripts/UnitStats.cs
@@ -3,5 +3,5 @@ using UnityEngine;
 public abstract class UnitStats : MonoBehaviour
 {
     public abstract Team Team { get; set; }
-    public abstract int Cost { get; }
+    public abstract float Cost { get; }
 }
