+++ Assets/Scripts/Units/FighterStats.cs
@@ -8,7 +8,7 @@ public class FighterStats : UnitStats, IUnit
 
 
     [Header("Attributes")]
-    [SerializeField] private int cost = 50;
+    [SerializeField] private float cost = 50;
     [SerializeField] private int maxHealth = 200;
     [SerializeField] private int currentHealth;
     [SerializeField] private int attackDamage = 20;
@@ -19,7 +19,7 @@ public class FighterStats : UnitStats, IUnit
 
 
     public override Team Team { get; set; }
-    public override int Cost => cost;
+    public override float Cost => cost;
     public GameObject GetGameObject() => gameObject;
     public Team GetTeam() => Team;
     public float GetAttackRange() => attackRange;
